% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_df.R
\name{arg_df}
\alias{arg_df}
\title{Returns initially processed data.frame from the export from the HDXExaminer}
\usage{
arg_df(filepath)
}
\arguments{
\item{filepath}{input file location}
}
\value{
Data.frame for further processing
}
\description{
Function used as internal function
}
