% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_prep.R
\name{arg_UN_FD}
\alias{arg_UN_FD}
\title{Returns initially processed data.frame from the export from the HDXExaminer}
\usage{
arg_UN_FD(filepath)
}
\arguments{
\item{filepath}{input file location}
}
\value{
Data.frame for further processing
}
\description{
Function used as internal function
}
