% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_call.R
\name{arguments_call}
\alias{arguments_call}
\title{Returns default arguments for the output_tp functions}
\usage{
arguments_call(filepath)
}
\arguments{
\item{filepath}{input file location}
}
\value{
The default arguments to output_tp functions.
}
\description{
Function used as internal function
}
