% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_call.R
\name{arguments_call2}
\alias{arguments_call2}
\title{Returns default arguments for the output_tp functions. Deut.Time}
\usage{
arguments_call2(filepath, states)
}
\arguments{
\item{filepath}{input file location}

\item{states}{states used}
}
\value{
The default arguments to output_tp functions.
}
\description{
Function used as internal function
}
