% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments_call.R
\name{arguments_call3}
\alias{arguments_call3}
\title{Returns default arguments for the output_tp functions. # replicates}
\usage{
arguments_call3(filepath, states, times)
}
\arguments{
\item{filepath}{input file location}

\item{states}{states used}

\item{times}{deuteration times}
}
\value{
The default arguments to output_tp functions.
}
\description{
Function used as internal function
}
