% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ave_tp.R
\name{av_tp}
\alias{av_tp}
\title{Preparatory function for average plot}
\usage{
av_tp(df, cola)
}
\arguments{
\item{df}{output from functions output_tp or output_tp or output_tp_proc.}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
plots of averages
}
\description{
Returns plots with average deuteration at each peptide.
}
