% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_sd_timecourses.R
\name{average_timecourse}
\alias{average_timecourse}
\title{Calculates average for time course data.}
\usage{
average_timecourse(filepath)
}
\arguments{
\item{filepath}{filepath to the All_results input file.}
}
\value{
data frame with average deuteration uptake data.
}
\description{
Calculates average for time course data.
}
