% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{boxplot_tp}
\alias{boxplot_tp}
\title{Plots boxplots for all the averages in the set}
\usage{
boxplot_tp(df, replicates = 3, ...)
}
\arguments{
\item{df}{average data frame. Generated using ave_timepoint() function.}

\item{replicates}{number of replicates in sample. Default set to 3.}

\item{...}{inherited boxplot parameters}
}
\value{
boxplots for average deuterium uptake per set.
}
\description{
Returns boxplots to compare sets between each other
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp(file_nm)
boxplot_tp(df=a, replicates=3)
}
