% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_tc.R
\name{CI_tc}
\alias{CI_tc}
\title{Critial interval calculation two sets of timecourses}
\usage{
CI_tc(sd_c, sd_v, replicates = 3, pv_cutoff = 0.01)
}
\arguments{
\item{sd_c}{dataframe of control}

\item{sd_v}{dataframe for variant}

\item{replicates}{number of replicates. Default set to 3.}

\item{pv_cutoff}{pvalue cutoff. Default set to 0.01}
}
\value{
Critical interval for 2 sets
}
\description{
Preparatory function for calculation of pvalue between sets.
}
