% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_pallettes.R
\name{color_ranges_Blue_Red_heat_map}
\alias{color_ranges_Blue_Red_heat_map}
\title{Returns color pallete from red to blue with number of colors for defined ranges}
\usage{
color_ranges_Blue_Red_heat_map(ranges, colors_initial)
}
\arguments{
\item{ranges}{vector of numbers. Should have the same mumber of positive and negative values and contain 0.}

\item{colors_initial}{additional color that should be first in the pallette.}
}
\value{
color scheme for number
}
\description{
Returns color pallete from red to blue with number of colors for defined ranges
}
\examples{
color_ranges_Blue_Red_heat_map(ranges=c(-Inf, -100, -50, 0, 50, 100, Inf), colors_initial="white")
}
