% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_functions_plot.R
\name{coverage_residue}
\alias{coverage_residue}
\title{Returns coverage per residue}
\usage{
coverage_residue(df1, start_col, end_col)
}
\arguments{
\item{df1}{output from functions output_tp or output_tp_proc.}

\item{start_col}{number of "Start" column in data.frame}

\item{end_col}{number of "Start" column in data.frame}
}
\value{
vector with coverage per residue
}
\description{
returns vector with coverage information
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp(file_nm)
coverage_residue(df1=a,start_col=2, end_col=3 )
}
