% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deuteration_woods_plots_frac.R
\name{deuteration_woods_timecourse_frac}
\alias{deuteration_woods_timecourse_frac}
\title{Return woods plots for the timecourse}
\usage{
deuteration_woods_timecourse_frac(
  input_data,
  states,
  replicates = 3,
  ylim = c(0, 1.2),
  ...
)
}
\arguments{
\item{input_data}{output from function output_tc(..., percent=TRUE)}

\item{states}{states, if missing all states used}

\item{replicates}{replicates}

\item{ylim}{y axis limits}

\item{...}{other parameters}
}
\value{
Woods plots for the timecourse
}
\description{
All the peptides are plotted based on their uptake.
}
