% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deuteration_woods_plots_frac.R
\name{deuteration_woods_timepoints_frac_list}
\alias{deuteration_woods_timepoints_frac_list}
\title{Return woods plots for the timepoints}
\usage{
deuteration_woods_timepoints_frac_list(
  list_input_data,
  times,
  replicates = 3,
  cola = NA,
  ylim = c(0, 1.1),
  ...
)
}
\arguments{
\item{list_input_data}{list of multiple outputs from function
output_tp(..., percent=TRUE)}

\item{times}{Deuteration times, if missing all deuteration times used}

\item{replicates}{replicates}

\item{cola}{colors, default NA}

\item{ylim}{y axis limits}

\item{...}{other parameters}
}
\value{
Woods plots for the timepoints
}
\description{
All the peptides are plotted based on their uptake.
}
