% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_tp_proc.R
\name{dif_tp_proc}
\alias{dif_tp_proc}
\title{Preparatory function for difference plot}
\usage{
dif_tp_proc(df, cola)
}
\arguments{
\item{df}{output from functions output_tp or output_tp_proc.}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
plots of difference in average
}
\description{
Returns plots with difference deuteration at each peptide.
}
