% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woods_uptake_plots_internal_functions.R
\name{duplicate_sets}
\alias{duplicate_sets}
\title{Duplicate set function}
\usage{
duplicate_sets(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
duplicate sets
}
\description{
Internal function
}
