% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extreme_input_prep_undeut.R
\name{extreme_input_undeut}
\alias{extreme_input_undeut}
\title{Makes input for Extreme for bimodal analysis.}
\usage{
extreme_input_undeut(hm_dir, replicates, timepoints)
}
\arguments{
\item{hm_dir}{directory in which all the folders which needs to be processed are}

\item{replicates}{number of replicates in sample}

\item{timepoints}{lists timepoints used in experiments.}
}
\value{
Inputs for extreme for all data prepared.
}
\description{
If data is missing it returns non-deuterated data in these columns.
}
\examples{
# extreme_input_undeut(hm_dir ="filepath", replicates = 2, timepoints =c(3, 60, 1800, 72000))
}
