% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_info_functions.R
\name{general_info}
\alias{general_info}
\title{Provides summary table for all data.sets.}
\usage{
general_info(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
Returns summary table.
}
\description{
Returns data frame sumamrizing general information about the data sets.
Function returns: Protein states, timepoints, number of replicates,  # peptides, \% coveregae, average peptide length and redundancy.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- general_info(file_nm)
}
