% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getCoords1}
\alias{getCoords1}
\title{function from plotfunctions package}
\usage{
getCoords1(pos = 1.1, side = 1, input = "p")
}
\arguments{
\item{pos}{position}

\item{side}{side of plot}

\item{input}{plot or figure position}
}
\value{
coordinates of margins
}
\description{
Margin coordinates
}
