% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heat_map_tc.R
\name{heat_map_tc}
\alias{heat_map_tc}
\title{Plots heat maps for time courses.}
\usage{
heat_map_tc(df, ranges = c(seq(0, 100, by = 10), Inf))
}
\arguments{
\item{df}{timecourse input}

\item{ranges}{ranges for coloring scheme. Default set to c(seq(0, 100, by=10), Inf)}
}
\value{
heat map for timecourses
}
\description{
Returns heat map on timecourses with raw data.
}
