% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heat_map_tp.R
\name{heat_map_tp}
\alias{heat_map_tp}
\title{Preparatory function for heat map}
\usage{
heat_map_tp(
  df,
  pv,
  sd,
  ranges = c(-Inf, seq(-30, 30, by = 10), Inf),
  pv_cutoff = 0.01,
  replicates = 3
)
}
\arguments{
\item{df}{average data frame. Generated using ave_timepoint() function.}

\item{pv}{pvalues dataframes calculated using pv_timepoint() function}

\item{sd}{standard deviation data.frame generated using sd_timepoint function}

\item{ranges}{ranges for coloring scheme. Default set to c(-Inf, seq(-30, 30, by=10), Inf)}

\item{pv_cutoff}{p-value cutoff here set up to 0.01}

\item{replicates}{number of replicates in sample. Default set to 3.}
}
\value{
heat map for timepoints
}
\description{
Returns heat map
}
