% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nan.data.frame.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Checks for NaN is data.frame}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{Data frame to be checked for NaN}
}
\value{
logical. Returns info if data.frame contains NaNs.
}
\description{
Function by Hong Ooi; https://stackoverflow.com/questions/18142117/how-to-replace-nan-value-with-zero-in-a-huge-data-frame
}
\examples{
## this function will overwrite the is.nan function that works only on vectors and matrices
df<-data.frame(c(0,NaN), c(1, 2))
is.nan(df)
df[is.nan(df)]<- 0
}
