% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_tp.R
\name{lab_dif}
\alias{lab_dif}
\title{Legend for difference in averages plot.}
\usage{
lab_dif(df, cola)
}
\arguments{
\item{df}{output from functions average difference}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
legend for difference in average plot for time points
}
\description{
Returns legend for difference in average plots. Preparatory function.
}
