% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_tp_proc.R
\name{lab_dif_proc}
\alias{lab_dif_proc}
\title{Preparatory function for difference plot for procent deuteration}
\usage{
lab_dif_proc(df, cola)
}
\arguments{
\item{df}{output from functions output_tp or output_tp_proc.}

\item{cola}{color pallette for different Protein States. As default Paired pallette from RColorBrewer is used.}
}
\value{
legends for procent deuteration plots
}
\description{
Returns legends for plots procent deuteration at each peptide.
}
