% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_volcano.R
\name{lab_vol}
\alias{lab_vol}
\title{Preparatory function for volcano plot legends}
\usage{
lab_vol(df, cola)
}
\arguments{
\item{df}{output from functions output_tp}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
legends for volcano plots
}
\description{
Returns volcano plots
}
