% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette_heatmap_legend.R
\name{legend_heat_map}
\alias{legend_heat_map}
\title{Legend for the heatmaps prep function.}
\usage{
legend_heat_map(ranges = c(-Inf, seq(-30, 30, by = 10), Inf))
}
\arguments{
\item{ranges}{ranges that are to be colored in the legend. Default ranges=c(-Inf,seq(-30, 30, by=10), Inf )}
}
\value{
legend for the heatmap
}
\description{
Returns names for legend for the heatmaps
}
