% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_heat_map_tc.R
\name{legend_heat_map_tc}
\alias{legend_heat_map_tc}
\title{Legend for the heatmaps for timecourses.}
\usage{
legend_heat_map_tc(df)
}
\arguments{
\item{df}{generated using output_tcourse()}
}
\value{
legend for the heatmap
}
\description{
Returns names for legend for the heatmaps. Extracts names from data.frame
}
