% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_heat_map_tc.R
\name{legend_heat_map_timecourse}
\alias{legend_heat_map_timecourse}
\title{Legend for the heatmaps prep function for timecourses.}
\usage{
legend_heat_map_timecourse(ranges = c(-Inf, seq(0, 100, by = 10), Inf))
}
\arguments{
\item{ranges}{ranges that are to be colored in the legend. Default ranges=c(-Inf,seq(-30, 30, by=10), Inf )}
}
\value{
legend for the heatmap
}
\description{
Returns names for legend for the heatmaps
}
