% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette_heatmap_legend.R
\name{legend_heat_map_tp}
\alias{legend_heat_map_tp}
\title{Legend for the heatmaps.Extracts names from data.frame}
\usage{
legend_heat_map_tp(df)
}
\arguments{
\item{df}{average data frame. Generated using ave_timepoint() function.}
}
\value{
legend for the heatmap
}
\description{
Returns names for legend for the heatmaps
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp(file_nm)
legend_heat_map_tp(df=a)
}
