% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette_heatmap_legend.R
\name{legend_heat_map_tp_proc}
\alias{legend_heat_map_tp_proc}
\title{Legend for the heatmaps percent.Extracts names from data.frame}
\usage{
legend_heat_map_tp_proc(df)
}
\arguments{
\item{df}{average data frame.}
}
\value{
legend for the heatmap prercent
}
\description{
Returns names for legend for the heatmaps
}
