% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woods_uptake_plots_internal_functions.R
\name{legend_nm_bottom}
\alias{legend_nm_bottom}
\title{Legend, bottom of the plots}
\usage{
legend_nm_bottom(names, cols)
}
\arguments{
\item{names}{labels}

\item{cols}{colors}
}
\value{
legend at the bottom of the plot
}
\description{
Internal function
}
