% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ave_tp.R
\name{legend_raw_ave}
\alias{legend_raw_ave}
\title{Legend for average plot.}
\usage{
legend_raw_ave(df, cola)
}
\arguments{
\item{df}{output from functions output_tp or output_tp_proc.}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
legend for average plot for time points
}
\description{
Returns legend with average plots. Preparatory function.
}
