% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_timepoint_proc_plots.R
\name{legend_raw_ave_proc}
\alias{legend_raw_ave_proc}
\title{Preparatory function to draw legends for average procent}
\usage{
legend_raw_ave_proc(df, cola)
}
\arguments{
\item{df}{output from functions output_tp or output_tp_proc.}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
legend for average deuteration procent for timepoints
}
\description{
Returns legend with average procent deuteration at each peptide.
}
