% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_timecourse_plots.R
\name{legend_raw_ave_tc}
\alias{legend_raw_ave_tc}
\title{Legend for average deuteration plot for timecourse.}
\usage{
legend_raw_ave_tc(df, cola)
}
\arguments{
\item{df}{output from functions output_tp or output_tp_proc.}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
legend for average plot for time course
}
\description{
Returns legend with average plots. Preparatory function.
}
