% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette_heatmap_legend.R
\name{legend_sig_peptides}
\alias{legend_sig_peptides}
\title{Legend for the significant peptides}
\usage{
legend_sig_peptides(ranges = c(-Inf, seq(-30, 30, by = 10), Inf))
}
\arguments{
\item{ranges}{ranges that are to be colored in the legend. Default ranges=c(-Inf,seq(-30, 30, by=10), Inf )}
}
\value{
legend for the heatmap
}
\description{
Returns names for legend for the significant peptides plots.
}
