% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woods_uptake_plots_internal_functions.R
\name{legend_states_PerD_bottom}
\alias{legend_states_PerD_bottom}
\title{Legend, bottom of the plots}
\usage{
legend_states_PerD_bottom(df, cols)
}
\arguments{
\item{df}{dataframe}

\item{cols}{colors}
}
\value{
legend at the bottom of the plot
}
\description{
Internal function
}
