% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_tc_bottom.R
\name{legend_tc_bottom}
\alias{legend_tc_bottom}
\title{Preparatory function returns legends for the timecourses.}
\usage{
legend_tc_bottom(df, cols)
}
\arguments{
\item{df}{data frame from which names will be extracted}

\item{cols}{colors to be used in legend}
}
\value{
legend at the bottom of the plot
}
\description{
Preparatory function
}
