% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_exchangeable_protons.R
\name{nb_exch_deut}
\alias{nb_exch_deut}
\title{Number of exchangeable protons}
\usage{
nb_exch_deut(df)
}
\arguments{
\item{df}{standard deviation from one sample}
}
\value{
vector with number of exchangeable protons
}
\description{
Provides a vector with number of exchangeable protons, calculated from the input table.
Number of protons calculated as peptide_length - 2 - number of Prolines in the peptide
that are not in the first position
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp(file_nm)
nb_exch_deut(a)
}
