% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_states.R
\name{nm_states}
\alias{nm_states}
\title{Lists names of states in data sets}
\usage{
nm_states(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
list of Protein States.
}
\description{
Returns vector with name of states used for choosing states for input functions generation.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
names_states<- nm_states(file_nm)
}
