% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_UD_FD.R
\name{output_FD}
\alias{output_FD}
\title{Prepares output for HDX-MS Full deuteration data}
\usage{
output_FD(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
data frame with reorganized data where in columns is uptake data for Protein States.
}
\description{
Returns a data frame for Full deuteration set
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<-output_FD(file_nm)

}
