% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_UD_FD.R
\name{output_FD_proc}
\alias{output_FD_proc}
\title{Prepares output for HDX-MS Full deuteration data  for procent deuteration.}
\usage{
output_FD_proc(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
data frame with reorganized data where in columns is procent deuteration for Protein States.
}
\description{
Returns a data frame for Full deuteration set
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_FD_proc(file_nm)
}
