% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tc_noSeq.R
\name{output_tcourse_noSeq_proc}
\alias{output_tcourse_noSeq_proc}
\title{Prepares output for HDX-MS for the timecourses without matched peptide sequences.}
\usage{
output_tcourse_noSeq_proc(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
data frame with reorganized data where in columns is procent deuteration data for Protein States.
}
\description{
Returns a dataframe organized for additional analysis.
Function does not match for the sequence. It can be used to match mutant sequences.
In columns are procent deuteration data for loaded protein states.
Number of replicates must be the same for all protein states.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tcourse_noSeq_proc(file_nm)
}
