% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tcourse.R
\name{output_tcourse_proc}
\alias{output_tcourse_proc}
\title{Prepares output for procent deuteration of HDX-MS for the timecourses.}
\usage{
output_tcourse_proc(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
Returns reprocessed data. Data.frame with reorganized data where in columns are deuterium uptake for different timepoints in the analysis.
}
\description{
Returns a dataframe organized for additional analysis.
In reprocessed data frame columns have time course data for procent deuteration.
Function takes raw data from HDXexaminer and organize data into columns with \% deuteration
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tcourse_proc(file_nm)
}
