% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecourse_output_subset.R
\name{output_tcourse_state}
\alias{output_tcourse_state}
\title{Prepares output for HDX-MS for the timecourses for deuteration uptake data for selected states.}
\usage{
output_tcourse_state(filepath, state)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}

\item{state}{list of states that will be chosen for analysis}
}
\value{
Returns reprocessed data. Data.frame with reorganized data where in columns are deuterium uptake for different timepoints in the analysis.
}
\description{
Returns a dataframe organized for additional analysis.
In reprocessed data frame columns have time course data for selected Protein sates.
Function takes raw data from HDXexaminer and organize data into columns with uptake deuteration data.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
names_states<- nm_states(file_nm)
summary<-output_tcourse_state(file_nm, state=names_states)
}
