% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tp.R
\name{output_tp}
\alias{output_tp}
\title{Prepares output for HDX-MS for the timepoints.}
\usage{
output_tp(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
data frame with reorganized data where in columns is uptake data for Protein States.
}
\description{
Returns a dataframe organized for additional analysis.
In columns are uptake data for loaded protein states.
In the columns are uptake data for all protein state.
Number of replicates must be the same for all protein states.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp(file_nm)
}
