% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tp_csv.R
\name{output_tp_csv}
\alias{output_tp_csv}
\title{Prepares output for deuteration uptake of HDX-MS for the time points}
\usage{
output_tp_csv(filepath, output)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}

\item{output}{name of the output csv file}
}
\value{
Returns reprocessed data. Data.frame with reorganized data where in columns are deuteration uptake for different timepoints in the analysis.
}
\description{
In reprocessed data frame columns have time points data for uptake data.
Function takes raw data from HDXexaminer and organize data into columns with uptake of deuterons
}
\examples{
\donttest{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp_csv(file_nm, "output.csv")
}
}
