% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tp.R
\name{output_tp_proc}
\alias{output_tp_proc}
\title{Prepares output for HDX-MS for the timepoints.}
\usage{
output_tp_proc(filepath)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}
}
\value{
Returns reprocessed data. Data.frame with reorganized data where in columns is procent deuteration data for Protein States.
}
\description{
Returns a dataframe organized for additional analysis.
In columns are protein states from the loaded projects.
Function takes raw data from HDXexaminer and organize data into columns with \% deuteration
Number of replicates must be the same for all protein states.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp_proc(file_nm)
}
