% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tp_proc_def.R
\name{output_tp_proc_def}
\alias{output_tp_proc_def}
\title{Prepares output for HDX-MS for the percent deuteration for the timepoints.}
\usage{
output_tp_proc_def(
  filepath,
  replicates,
  states,
  times,
  seq_match = F,
  csv = "NA"
)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}

\item{replicates}{number of replicates to be used in analysis. The function takes number of replicates up to specified number. If no argument provided number maximal common number of replicates it used.}

\item{states}{function allows to choose what states should be used for analysis. Default all states are used.}

\item{times}{lists the deuteration times to be used in analysis. Default all states used.}

\item{seq_match}{Flag allows to choose if the peptide sequences should be matched between states. seq_match=F signifies no sequence matching, seq_match=T states that the sequences are matched between the sets.}

\item{csv}{Flag allowing saving the output as csv. With default csv="NA", data is not saved. If csv output is desided, provide output name.}
}
\value{
data frame with reorganized data where in columns is the percent deuteration for Protein States.
}
\description{
Returns a data frame organized for additional analysis.
In columns are uptake data for loaded protein states.
In the columns are uptake data for all protein state.
Number of replicates, protein states and Deut.Times can be specified
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
#a<- output_tp_proc_def(filepath=file_nm) ###all default parameters used
#a<-output_tp_proc_def(filepath=file_nm, replicates=3, states=c("Bound", "Unbound"), times=c("3.00s", "72000.00s"), seq_match=T, csv="NA")
}
