% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tp.R
\name{output_tp_states}
\alias{output_tp_states}
\title{Prepares output for HDX-MS for the timepoints at chosen protein states}
\usage{
output_tp_states(filepath, states)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}

\item{states}{list of states that will be chosen for analysis}
}
\value{
data frame with reorganized data where in columns is uptake data for Protein States.
}
\description{
Returns a dataframe organized for additional analysis.
In columns are uptake data for chosen protein states.
In the columns are uptake data for chosen protein states
Number of replicates must be the same for all protein states.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
names_states<- nm_states(file_nm)
a<- output_tp_states(file_nm, states=names_states)
}
