% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette_heatmap_legend.R
\name{pallette_legend}
\alias{pallette_legend}
\title{Color scheme using heatmap. Legend Extracts names from data.frame}
\usage{
pallette_legend(col_pallette)
}
\arguments{
\item{col_pallette}{pallette to be used in the heat map}
}
\value{
legend for the heatmap
}
\description{
Returns names for legend for the heatmaps
}
