% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette_heatmap_legend.R
\name{pallette_ll}
\alias{pallette_ll}
\title{Color scheme using heatmap. Legend extracts names from data frame}
\usage{
pallette_ll(pallette, lab)
}
\arguments{
\item{pallette}{pallette to be used in the heat map}

\item{lab}{labels to be used in pallette}
}
\value{
legend for the heatmap
}
\description{
Returns names for legend for the heatmaps
}
