% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woods_uptake_plots_internal_functions.R
\name{pl_gen_ch_frac}
\alias{pl_gen_ch_frac}
\title{Prepares the plot window for the woods functions}
\usage{
pl_gen_ch_frac(df, ddlab = 3, ...)
}
\arguments{
\item{df}{dataframe}

\item{ddlab}{label}

\item{...}{other}
}
\value{
Plot window
}
\description{
Internal function
}
