% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woods_uptake_plots_internal_functions.R
\name{pl_gen_ch_list_frac}
\alias{pl_gen_ch_list_frac}
\title{Prepares the plot window for the woods functions}
\usage{
pl_gen_ch_list_frac(min_x, max_x, ddlab = 3, ...)
}
\arguments{
\item{min_x}{x axis starting number}

\item{max_x}{x axis max value}

\item{ddlab}{label}

\item{...}{other}
}
\value{
Plot window
}
\description{
Internal function
}
