% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woods_uptake_plots_internal_functions.R
\name{pl_gen_uptake}
\alias{pl_gen_uptake}
\title{Prepares the plot window for the woods functions}
\usage{
pl_gen_uptake(df, timepoints, ddlab = 1, ...)
}
\arguments{
\item{df}{dataframe}

\item{timepoints}{deuteration times used}

\item{ddlab}{label}

\item{...}{other}
}
\value{
Plot window
}
\description{
Internal function
}
