% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions_2states.R
\name{plot_cor_pts_elps}
\alias{plot_cor_pts_elps}
\title{Correlation functions plots and prep data}
\usage{
plot_cor_pts_elps(
  df,
  replicates = 3,
  times,
  state1,
  state2,
  position_col = NA,
  length1 = NA,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{df}{input from tp function}

\item{replicates}{number of replicates in sample. Default set to 3.}

\item{times}{times to be used}

\item{state1}{First state}

\item{state2}{second state}

\item{position_col}{default NA, flag that allows to color based on domain,
if TRUE then the length1 vector needs to be provided}

\item{length1}{vector of numeric}

\item{xlim}{x-axis range}

\item{ylim}{y-axis range}

\item{...}{other plotting parameters}
}
\value{
Correlation plot
}
\description{
Correlation of two states, the st.dev are shown as ellipsoids.
}
