% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_functions_2states.R
\name{plot_correlation_functions}
\alias{plot_correlation_functions}
\title{Plot correlated states}
\usage{
plot_correlation_functions(df, replicates = 3, ranges, mar_x = 1)
}
\arguments{
\item{df}{standard deviation from one sample}

\item{replicates}{number of replicates. Default set to 3.}

\item{ranges}{ylim range and and how often ticks}

\item{mar_x}{x margin variable to be changed if needed}
}
\value{
plot for the correlation
}
\description{
Correlation of two states, the st.dev are shown as ellipsoids.
}
\examples{
sd1<-data.frame(c(0.1, 0.12, 0.13, 0.09, 0.11, 0.10))
CI_single(s1=sd1, replicates=3)
}
