% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppar.R
\name{ppar_bottom_legend}
\alias{ppar_bottom_legend}
\title{Preparation of figure window with area for figure at the bottom.}
\usage{
ppar_bottom_legend(mfrow2)
}
\arguments{
\item{mfrow2}{mfrow: number of Multiple Figures (use ROW-wise).}
}
\value{
modified par function with adjusted parameters
}
\description{
Prepares a plotting window with specified margins with specific number of figure row and columns.
}
\examples{
ppar_bottom_legend(c(2,3))
}
