% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppar.R
\name{ppar_wider}
\alias{ppar_wider}
\title{Preparation of figure window with more area on west side of plot.}
\usage{
ppar_wider(mfrow2)
}
\arguments{
\item{mfrow2}{mfrow: number of Multiple Figures (use ROW-wise).}
}
\value{
default plotting window
}
\description{
Prepares a plotting window with specified margins with specific number of figure row and columns.
}
\examples{
ppar_wider(c(2,1))
}
