% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppar.R
\name{pparLM}
\alias{pparLM}
\title{Preparation of figure window. small margins}
\usage{
pparLM(mfrow2)
}
\arguments{
\item{mfrow2}{mfrow: number of Multiple Figures (use ROW-wise).}
}
\value{
modified par function with adjusted parameters
}
\description{
Prepares a plotting window with specified margins with specific number of figure row and columns.
}
\examples{
pparLM(c(2,1))
}
