% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_timecourse_functions.R
\name{prep_timecourse_plot_ave}
\alias{prep_timecourse_plot_ave}
\title{Prepares function for plotting averages in timecourse}
\usage{
prep_timecourse_plot_ave(control_df, variant_df, replicates = 3)
}
\arguments{
\item{control_df}{dataframe of control}

\item{variant_df}{dataframe for variant}

\item{replicates}{number of replicates. Default set to 3.}
}
\value{
dataframes with matched peptides in time course
}
\description{
Preparatory function
}
