% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_timecourse_functions.R
\name{prep_timecourse_plot_sd}
\alias{prep_timecourse_plot_sd}
\title{Prepares function for Critical interval for timecourses}
\usage{
prep_timecourse_plot_sd(
  control_df_up,
  variant_df_up,
  replicates = 3,
  pv_cutoff = 0.01
)
}
\arguments{
\item{control_df_up}{dataframe of control}

\item{variant_df_up}{dataframe for variant}

\item{replicates}{number of replicates. Default set to 3.}

\item{pv_cutoff}{cut off of pvalue used in calculation of critical interval. Default set to 0.01}
}
\value{
Critial interval for all sets
}
\description{
Preparatory function
}
