% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_timecourse.R
\name{pv_timecourse}
\alias{pv_timecourse}
\title{pvalue calculation between two sets of the data at certain timepoint}
\usage{
pv_timecourse(df_c, df_v, replicates = 3)
}
\arguments{
\item{df_c}{dataframe of control}

\item{df_v}{dataframe for variant}

\item{replicates}{number of replicates. Default set to 3.}
}
\value{
pvalue comparisons between two sets.
}
\description{
Preparatory function for calculation of pvalue between sets.
}
