% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymol_str.R
\name{pymol_str}
\alias{pymol_str}
\title{Preparatory function writing pymol scripts}
\usage{
pymol_str(ind1)
}
\arguments{
\item{ind1}{vector of numbers (residues)}
}
\value{
string with + as a separator.
}
\description{
Function rearrange vector to string by adding + sign between the numbers.
}
\examples{
res<-c(1,5, 19, 100, 109)
pymol_str(res)
}
