% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr.cbind.na.R
\name{qpcr.cbind.na}
\alias{qpcr.cbind.na}
\title{Hidden function from qpcR package, typical usage as qpcR:::cbind.na}
\usage{
qpcr.cbind.na(..., deparse.level = 1)
}
\arguments{
\item{...}{vectors}

\item{deparse.level}{set to 1 as default}
}
\value{
data frame with NA
}
\description{
Combine data of unequal row length avoiding repetition or errors by filling with NAs. In contrast to classical cbind, cbind.na can be used to combine data such as
}
\examples{
qpcr.cbind.na(1:10, 1:3)

}
