% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_functions.R
\name{ranges_function_tc}
\alias{ranges_function_tc}
\title{Gives ranges for the averages for time course analysis}
\usage{
ranges_function_tc(df_ave, values_df)
}
\arguments{
\item{df_ave}{average per residues}

\item{values_df}{data frame with values.}
}
\value{
ranges per set
}
\description{
Function used as internal function to get ranges in the function.
}
