% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppar.R
\name{reset_par}
\alias{reset_par}
\title{Reset plotting window parameters to default}
\usage{
reset_par()
}
\value{
default plotting window parameters
}
\description{
function by Farid Cheraghi,
https://stackoverflow.com/questions/9292563/reset-the-graphical-parameters-back-to-default-values-without-use-of-dev-off
function resets plotting window parameters
}
\examples{
reset_par()
}
