% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_sd_timecourses.R
\name{sd_timecourse}
\alias{sd_timecourse}
\title{Returns standard deviation for uptake data for timecourses.}
\usage{
sd_timecourse(filepath)
}
\arguments{
\item{filepath}{filepath to the All_results input file.}
}
\value{
Data.frame with standard deviation.
}
\description{
Calculates standard deviation for timecourse data.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
sd_timecourse(filepath=file_nm)
}
