% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave_sd_timecourse_proc.R
\name{sd_timecourse_proc}
\alias{sd_timecourse_proc}
\title{Returns standard deviation for percent deuteration data for timecourses.}
\usage{
sd_timecourse_proc(filepath)
}
\arguments{
\item{filepath}{filepath to the All_results input file.}
}
\value{
Data.frame with standard deviation.
}
\description{
Calculates standard deviation for time course data.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
sd_timecourse(filepath=file_nm)
}
