% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_ave_diff_pv_tp.R
\name{sd_timepoint}
\alias{sd_timepoint}
\title{Returns standard deviation for dataframe.}
\usage{
sd_timepoint(df, replicates = 3)
}
\arguments{
\item{df}{output from functions output_tp or output_tp_proc.}

\item{replicates}{number of replicates used. Default is set to replicates=3}
}
\value{
Data.frame with standard deviation.
}
\description{
Calculates standard deviation for the number of replicates in the function.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tp(file_nm)
sd<-sd_timepoint(df=a, replicates=3)
}
