% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_functions_plot.R
\name{significant_peptide_uptake}
\alias{significant_peptide_uptake}
\title{Function returns which peptides are significantly based of pv_cutoff and Critial interval}
\usage{
significant_peptide_uptake(df_av, pv, sd, pv_cutoff = 0.01, replicates = 3)
}
\arguments{
\item{df_av}{data.frame with averages created using ave_timepoint() function}

\item{pv}{data.frame with pvalues created using pv_timepoint() function}

\item{sd}{data.frame with standard deviations created using sd_timepoint() function}

\item{pv_cutoff}{cuttoff for Critical interval. Default=0.01}

\item{replicates}{number of replicates as default set to 3.}
}
\value{
ranges per set
}
\description{
Returns data frame with significant peptides.
}
