% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_CI_sdev.R
\name{summary_sd_CI}
\alias{summary_sd_CI}
\title{Provides summary table with Critical interval and standard deviation within the set.}
\usage{
summary_sd_CI(filepath, replicates = 3)
}
\arguments{
\item{filepath}{filepath to the input file. Input file is All_results table from HDX_Examiner, where all the fields are marked for export.}

\item{replicates}{number of replicates. Default set to 3.}
}
\value{
Returns summary table.
}
\description{
Returns summary data.
Function returns: Protein states, timepoints, number of replicates,  # peptides, \% coveregae, average peptide length and redundancy.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- summary_sd_CI(file_nm, replicates=3)
}
