% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoint_output_subset.R
\name{timepoint_output_subset}
\alias{timepoint_output_subset}
\title{Returns data.frame with averages from analysis for uptake file, standard deviation, p-values for selected protein states.}
\usage{
timepoint_output_subset(filepath, states)
}
\arguments{
\item{filepath}{path to All.Data.csv input from HDX-Examiner.}

\item{states}{list of states that will be chosen for analysis}
}
\value{
data frame with analysis for uptake file, standard deviation, p-values for selected protein states.
}
\description{
Returns information from analysis data.frame for selected protein states.
Sets are compared to the first state in the input file.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
names_states<- nm_states(file_nm)
summary<-timepoint_output_subset(file_nm, states=names_states)
}
