% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_output_tp_csv.R
\name{verbose_timepoint_output}
\alias{verbose_timepoint_output}
\title{Returns csv with averages from analysis for uptake file, standard deviation, p-values.}
\usage{
verbose_timepoint_output(filepath, output_name)
}
\arguments{
\item{filepath}{path to All.Data.csv input from HDX-Examiner.}

\item{output_name}{name of the output in csv format.}
}
\value{
csv with analysis for uptake file, standard deviation, p-values for all protein states.
}
\description{
Returns information from analysis and save it as csv file.
Sets are compared to the first state in the input file.
}
\examples{
\donttest{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
verbose_timepoint_output(file_nm, "output_file_name.csv")
}
}
