% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_output_tp_csv.R
\name{verbose_timepoint_output_subset}
\alias{verbose_timepoint_output_subset}
\title{Returns csv with averages from analysis for uptake file, standard deviation, p-values for selected protein states.}
\usage{
verbose_timepoint_output_subset(filepath, states, output_name)
}
\arguments{
\item{filepath}{path to All.Data.csv input from HDX-Examiner.}

\item{states}{list of states that will be chosen for analysis}

\item{output_name}{name of the output in csv format.}
}
\value{
csv with analysis for uptake file, standard deviation, p-values for selected protein states.
}
\description{
Returns information from analysis and save it as csv file for selected protein states.
Sets are compared to the first state in the input file.
}
\examples{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
names_states<- nm_states(file_nm)
verbose_timepoint_output_subset(file_nm, "output_file_name.csv", states=names_states)
}
