% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_volcano.R
\name{vol_tp}
\alias{vol_tp}
\title{Preparatory function for volcano plot}
\usage{
vol_tp(df1, pv, CI, pv_cutoff = 0.01, cola)
}
\arguments{
\item{df1}{differences in averages data.frame calculated using diff_ave function}

\item{pv}{pvalues dataframes calculated using pv_timepoint function}

\item{CI}{critical interval, here is multiple sets are using maximun CI is used.}

\item{pv_cutoff}{p-value cutoff here set up to 0.01}

\item{cola}{color pallette for different Protein States. As default Paired pallette from color.Brewer is used.}
}
\value{
volcano plots
}
\description{
Returns volcano plots
}
