% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptake_plots.R
\name{uptake_plots}
\alias{uptake_plots}
\title{Uptake plots}
\usage{
uptake_plots(
  input_data,
  timepoints,
  replicates = 3,
  cola = NA,
  seq_match = TRUE
)
}
\arguments{
\item{input_data}{output from function output_tp(..., percent=T)}

\item{timepoints}{the labeling times}

\item{replicates}{replicates}

\item{cola}{colors, default NA}

\item{seq_match}{Flag TRUE or FALSE, default TRUE, match sequence of the protein states}
}
\value{
Uptake plots
}
\description{
Uptake plots per peptide
}
\examples{
\donttest{
file_nm<-system.file("extdata", "All_results_table.csv", package = "HDXBoxeR")
a<- output_tc(file_nm, percent=TRUE)
x=c(3,60, 1800, 72000)
uptake_plots(a, x)
}
}
